renaming items might move them between disks, which could be messing up local flags, which causes referrals to invalid blocks.
Gonna remove the Option from InodeLocation, and automatically remove it when the inode is being written/read. This means we can also remove the OnThisDisk flag, since we will always know where it is.

In fact, might as well remove all of the local stuff. On read/write we will deduce its location fully.

InodeLocations are now smarter and can self detect if they require a disk number

9 Failing tests. Looks like a few of them are related to virtual disks not being created.